# restfull/FEMS/swagger.yaml
openapi: 3.0.3
info:
  title: TZW LTD Fire Extinguisher Management API
  version: 1.0.0
  description: |
    Consolidated API documentation for the FEMS backend services behind the API gateway.

    Covered services:
    - Authentication and user administration
    - Facility and customer management
    - Fire extinguisher registration and tracking
    - Inspection scheduling and recording
    - Maintenance logging
    - Notifications and escalations
    - Reporting and audit access
servers:
  - url: http://localhost:5000/api
    description: Local API gateway
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit10Param:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Inspection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        extinguisherId:
          type: string
          format: uuid
        extinguisherCode:
          type: string
        inspectorId:
          type: string
          format: uuid
        inspectorName:
          type: string
        inspectionDate:
          type: string
          format: date
        inspectionTime:
          type: string
          nullable: true
          example: '10:30'
        findings:
          type: string
          nullable: true
        status:
          type: string
          enum: [Scheduled, Completed, Requires Service, Failed]
        nextInspectionDate:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
    InspectionCreateRequest:
      type: object
      required: [extinguisherId, inspectorId, inspectionDate, status]
      properties:
        extinguisherId:
          type: string
          format: uuid
        inspectorId:
          type: string
          format: uuid
        inspectionDate:
          type: string
          format: date
        inspectionTime:
          type: string
          nullable: true
          example: '10:30'
        findings:
          type: string
          nullable: true
        status:
          type: string
          enum: [Scheduled, Completed, Requires Service, Failed]
        nextInspectionDate:
          type: string
          format: date
          nullable: true
...
security:
  - bearerAuth: []
paths:
...
  /inspections:
    post:
      tags: [Inspections]
      summary: Schedule or record an inspection
      description: |
        Users can schedule only.
        Admins and inspectors can also record completed or failed outcomes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionCreateRequest'
      responses:
        '201':
          description: Inspection scheduled or recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  data:
                    $ref: '#/components/schemas/Inspection'
        '403':
          description: User role cannot submit the requested outcome
    get:
      tags: [Inspections]
      summary: List inspections
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/Limit10Param'
        - name: extinguisherId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [Scheduled, Completed, Requires Service, Failed]
      responses:
        '200':
          description: Inspection list
  /inspections/{id}:
    get:
      tags: [Inspections]
      summary: Get an inspection by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Inspection details
    put:
      tags: [Inspections]
      summary: Update an inspection by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionCreateRequest'
      responses:
        '200':
          description: Inspection updated
...
  /reports/inspections:
    get:
      tags: [Reports]
      summary: Inspection report
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv, pdf]
      responses:
        '200':
          description: Report payload or downloadable export
...